// src/app/interfaces/user.ts
export interface User {
    iduser: number;
    name: string;
    surname: string;
    email: string;
    birth_date: string;
    gender: string;
    user: string;
    password: string;
    subtitle: string;
    bio: string;
    profile_picture: string;
    header_picture: string;
    instagram: string;
    facebook: string;
    twitter: string;
    register_date: string;
}